requirements.txt
pytest
playwright
Appium-Python-Client
allure-pytest